ReStructured folder for optimizing next prod build
