+++ currency-frontend/__tests__/index.spec.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import '@testing-library/jest-dom';
 import { render, screen } from '@testing-library/react';
 import Index from '../pages/index'
-import { WELCOME_MESSAGE } from '../pages/constants/labels'
+import { WELCOME_MESSAGE } from '../constants/labels'
 
 describe('Pages', () => {
   describe('Index', () => {
