+++ currency-frontend/actions/currenciesAction.ts
