+++ currency-frontend/common/baseCurrencySelector/baseCurrencySelectorComponent.tsx
