+++ currency-frontend/common/currencyExchangeInfo/currencyExchangeInfoComponent.tsx
