+++ currency-frontend/common/currencyInput/currencyInputComponent.tsx
