+++ currency-frontend/common/navigation/navigationComponent.tsx
