+++ currency-frontend/constants/actionTypes.ts
