+++ currency-frontend/constants/errorMessages.ts
