+++ currency-frontend/constants/formats.ts
