+++ currency-frontend/constants/labels.ts
