+++ currency-frontend/constants/pagePaths.ts
