+++ currency-frontend/helpers/componentDataParser.ts
