+++ currency-frontend/helpers/openExchangeResponseParser.ts
