+++ currency-frontend/helpers/urlManipulator.ts
