+++ currency-frontend/pages/dashboard.tsx
@@ -1,6 +1,6 @@
 import HistoricalExchangeRatesComponent from "./modules/dashboard/historicalExchangeRates/historicalExchangeRatesComponent";
 import LiveExchangeRatesComponent from "./modules/dashboard/liveExchangeRates/liveExchangeRatesComponent";
-import Navigation from "./common/navigation/navigationComponent";
+import Navigation from "../common/navigation/navigationComponent";
 import CurrencySelectorComponent from "./modules/currencySelector/currencySelectorComponent";
 
 const Dashboard:React.FC = ()=>{
