+++ currency-frontend/reducers/currenciesReducer.ts
