+++ currency-frontend/reducers/historicalExchangeRateReducer.ts
