+++ currency-frontend/reducers/index.ts
