+++ currency-frontend/types/next.ts
