+++ currency-frontend/types/openExchangeResponseTypes.ts
