+++ currency-frontend/types/openExchnageRequestTypes.ts
