+++ currency-frontend/types/stateTypes.ts
